% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoefExplainer.R
\name{CoefExplainer}
\alias{CoefExplainer}
\title{Explore the Meaning of Coefficients in a Linear Model with Categorical Covariates}
\usage{
CoefExplainer(data, formula)
}
\arguments{
\item{data}{a data.frame that contains the covariates of the \code{formula} as
columns.}

\item{formula}{a standard R formula (e.g. \code{y ~ group + batch + sex}). Unlike
most R functions, the covarites in the formula (e.g. \code{group}, \code{batch} etc.)
can only reference columns of \code{data}.}
}
\value{
A list with stuff.
}
\description{
The function takes a data frame and a formula. It fits a linear model and parses
the formula so that the coefficients can be visuallized easily.
}
\details{
Note: the function can only handle categorical covariates, because
continuous variables are more difficult to visualize.
}
\examples{

  peng <- palmerpenguins::penguins
  # Remove any NA's
  peng <- peng[! apply(peng, 1, function(row) any(is.na(row))), ]
  peng


  coefExplFit <- CoefExplainer(peng, flipper_length_mm ~ species + island + sex)

  plotModel(coefExplFit)

  plotModelMatrix(coefExplFit)

  plotCoef(coefExplFit)

}
