% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoefExplainer.R
\name{get_rhs}
\alias{get_rhs}
\title{formulatools::rhs.vars is problematic if a term only appears in the
interaction term...}
\usage{
get_rhs(formula, data)
}
\description{
formulatools::rhs.vars is problematic if a term only appears in the
interaction term...
}
