% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotAll}
\alias{plotAll}
\title{Make a combined plot of Model, ModelMatrix, and Coef}
\usage{
plotAll(coefExplFit, title = as.character(coefExplFit$formula))
}
\arguments{
\item{coefExplFit}{the result from the \code{\link{CoefExplainer}()}.}

\item{title}{the title of the plot. By default the formula}
}
\description{
Make a combined plot of Model, ModelMatrix, and Coef
}
