% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotCoef}
\alias{plotCoef}
\title{Make a bar plot of the coefficients}
\usage{
plotCoef(coefExplFit, interceptSeparate = TRUE)
}
\arguments{
\item{coefExplFit}{the result from the \code{\link{CoefExplainer}()}.}

\item{interceptSeparate}{should the intercept be placed separately}
}
\description{
Make a bar plot of the coefficients
}
