% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotModel}
\alias{plotModel}
\title{Plot each Group of Datapoints and the Coefficients}
\usage{
plotModel(coefExplFit, showCoefficients = TRUE)
}
\arguments{
\item{coefExplFit}{the result from the \code{\link{CoefExplainer}()}.}

\item{showCoefficients}{a boolean that decides if the coefficients are
plotted above the data.}
}
\description{
The groups consists of the different combinations of values for the
covariates on the right hand side of the formula.
}
