% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotModelMatrix}
\alias{plotModelMatrix}
\title{Visualize each combination of covariates of the model matrix}
\usage{
plotModelMatrix(coefExplFit)
}
\arguments{
\item{coefExplFit}{the result from the \code{\link{CoefExplainer}()}.}
}
\description{
Visualize each combination of covariates of the model matrix
}
